CBP - Stocks Transactions - Cash Money Destock  Fix GetManager- QA #3026
