+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-cash-money-destock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/cash_money_destock/developer/bitdubai/version_1/StockTransactionsCashMoneyDestockPluginRoot.java
@@ -5,6 +5,7 @@ import com.bitdubai.fermat_api.FermatException;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.abstract_classes.AbstractPlugin;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededAddonReference;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededPluginReference;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginVersionReference;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DatabaseManagerForDevelopers;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabase;
@@ -51,7 +52,6 @@ import java.util.UUID;
  * Created by franklin on 16/11/15.
  */
 public class StockTransactionsCashMoneyDestockPluginRoot extends AbstractPlugin  implements
-        CashMoneyDestockManager,
         DatabaseManagerForDevelopers {
 
     public StockTransactionsCashMoneyDestockPluginRoot() {
@@ -81,13 +81,13 @@ public class StockTransactionsCashMoneyDestockPluginRoot extends AbstractPlugin 
 
     @Override
     public void start() throws CantStartPluginException {
-        stockTransactionCashMoneyDestockManager = new StockTransactionCashMoneyDestockManager(pluginDatabaseSystem, pluginId);
-        try {
+
+       try {
             Database database = pluginDatabaseSystem.openDatabase(pluginId, StockTransactionsCashMoneyDestockDatabaseConstants.CASH_MONEY_DESTOCK_DATABASE_NAME);
 
             //Buscar la manera de arrancar el agente solo cuando hayan transacciones diferentes a COMPLETED
             System.out.println("******* Init Cash Money Destock ******");
-            //testDestock();
+
             startMonitorAgent();
 
             database.closeDatabase();
@@ -116,6 +116,11 @@ public class StockTransactionsCashMoneyDestockPluginRoot extends AbstractPlugin 
     }
 
     @Override
+    public FermatManager getManager() {
+        return new StockTransactionCashMoneyDestockManager(pluginDatabaseSystem, pluginId);
+    }
+
+    @Override
     public List<DeveloperDatabase> getDatabaseList(DeveloperObjectFactory developerObjectFactory) {
         StockTransactionsCashMoneyDestockDeveloperFactory stockTransactionsCashMoneyDestockDeveloperFactory = new StockTransactionsCashMoneyDestockDeveloperFactory(pluginDatabaseSystem, pluginId);
         return stockTransactionsCashMoneyDestockDeveloperFactory.getDatabaseList(developerObjectFactory);
@@ -140,34 +145,6 @@ public class StockTransactionsCashMoneyDestockPluginRoot extends AbstractPlugin 
         return developerDatabaseTableRecordList;
     }
 
-    @Override
-    public void createTransactionDestock(String publicKeyActor, FiatCurrency fiatCurrency, String cbpWalletPublicKey, String cshWalletPublicKey, String cashReference, float amount, String memo, float priceReference, OriginTransaction originTransaction) throws CantCreateCashMoneyDestockException {
-        java.util.Date date = new java.util.Date();
-        Timestamp timestamp = new Timestamp(date.getTime());
-        CashMoneyDestockTransactionImpl cashMoneyRestockTransaction = new CashMoneyDestockTransactionImpl(
-                UUID.randomUUID(),
-                publicKeyActor,
-                fiatCurrency,
-                cbpWalletPublicKey,
-                cshWalletPublicKey,
-                memo,
-                "INIT TRANSACTION",
-                cashReference,
-                amount,
-                timestamp,
-                TransactionStatusRestockDestock.INIT_TRANSACTION,
-                priceReference,
-                originTransaction);
-
-        try {
-            stockTransactionCashMoneyDestockManager.saveCashMoneyDestockTransactionData(cashMoneyRestockTransaction);
-        } catch (DatabaseOperationException e) {
-            e.printStackTrace();
-        } catch (MissingCashMoneyDestockDataException e) {
-            e.printStackTrace();
-        }
-    }
-
     private StockTransactionsCashMoneyDestockMonitorAgent stockTransactionsCashMoneyDestockMonitorAgent;
     /**
      * This method will start the Monitor Agent that watches the asyncronic process registered in the bank money restock plugin
@@ -184,14 +161,6 @@ public class StockTransactionsCashMoneyDestockPluginRoot extends AbstractPlugin 
 
             stockTransactionsCashMoneyDestockMonitorAgent.start();
         }else stockTransactionsCashMoneyDestockMonitorAgent.start();
-    }
-
-    private void testDestock(){
-        try {
-            createTransactionDestock("publicKeyActor", FiatCurrency.VENEZUELAN_BOLIVAR, "cbpWalletPublicKey", "cshWalletPublicKey", "cashReference", 250, "memo", 250, OriginTransaction.STOCK_INITIAL);
-        } catch (CantCreateCashMoneyDestockException e) {
-            e.printStackTrace();
-        }
     }
 
 }
