+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-cash-money-destock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/cash_money_destock/developer/bitdubai/version_1/structure/StockTransactionCashMoneyDestockManager.java
@@ -1,13 +1,19 @@
 package com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.cash_money_destock.developer.bitdubai.version_1.structure;
 
+import com.bitdubai.fermat_api.layer.all_definition.enums.FiatCurrency;
 import com.bitdubai.fermat_api.layer.all_definition.exceptions.InvalidParameterException;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.DatabaseTableFilter;
 import com.bitdubai.fermat_api.layer.osa_android.database_system.PluginDatabaseSystem;
 import com.bitdubai.fermat_cbp_api.all_definition.business_transaction.CashMoneyTransaction;
+import com.bitdubai.fermat_cbp_api.all_definition.enums.OriginTransaction;
+import com.bitdubai.fermat_cbp_api.all_definition.enums.TransactionStatusRestockDestock;
+import com.bitdubai.fermat_cbp_api.layer.stock_transactions.cash_money_destock.exceptions.CantCreateCashMoneyDestockException;
+import com.bitdubai.fermat_cbp_api.layer.stock_transactions.cash_money_destock.interfaces.CashMoneyDestockManager;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.cash_money_destock.developer.bitdubai.version_1.database.StockTransactionsCashMoneyDestockDatabaseDao;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.cash_money_destock.developer.bitdubai.version_1.exceptions.DatabaseOperationException;
 import com.bitdubai.fermat_cbp_plugin.layer.stock_transactions.cash_money_destock.developer.bitdubai.version_1.exceptions.MissingCashMoneyDestockDataException;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.UUID;
 
@@ -17,7 +23,8 @@ import java.util.UUID;
  *
  * Created by franklin on 17/11/15.
  */
-public class StockTransactionCashMoneyDestockManager {
+public class StockTransactionCashMoneyDestockManager implements
+        CashMoneyDestockManager {
     private final PluginDatabaseSystem pluginDatabaseSystem;
     private final UUID pluginId;
 
@@ -33,12 +40,40 @@ public class StockTransactionCashMoneyDestockManager {
         this.pluginId             = pluginId            ;
     }
 
+    @Override
+    public void createTransactionDestock(String publicKeyActor, FiatCurrency fiatCurrency, String cbpWalletPublicKey, String cshWalletPublicKey, String cashReference, float amount, String memo, float priceReference, OriginTransaction originTransaction) throws CantCreateCashMoneyDestockException {
+        java.util.Date date = new java.util.Date();
+        Timestamp timestamp = new Timestamp(date.getTime());
+        CashMoneyDestockTransactionImpl cashMoneyRestockTransaction = new CashMoneyDestockTransactionImpl(
+                UUID.randomUUID(),
+                publicKeyActor,
+                fiatCurrency,
+                cbpWalletPublicKey,
+                cshWalletPublicKey,
+                memo,
+                "INIT TRANSACTION",
+                cashReference,
+                amount,
+                timestamp,
+                TransactionStatusRestockDestock.INIT_TRANSACTION,
+                priceReference,
+                originTransaction);
+
+        try {
+            saveCashMoneyDestockTransactionData(cashMoneyRestockTransaction);
+        } catch (DatabaseOperationException e) {
+            e.printStackTrace();
+        } catch (MissingCashMoneyDestockDataException e) {
+            e.printStackTrace();
+        }
+    }
+
     private StockTransactionsCashMoneyDestockDatabaseDao getStockTransactionCashMoneyDestockDao() {
 
         return new StockTransactionsCashMoneyDestockDatabaseDao(pluginDatabaseSystem, pluginId);
     }
 
-    public void saveCashMoneyDestockTransactionData(CashMoneyTransaction cashMoneyTransaction) throws DatabaseOperationException, MissingCashMoneyDestockDataException {
+    private void saveCashMoneyDestockTransactionData(CashMoneyTransaction cashMoneyTransaction) throws DatabaseOperationException, MissingCashMoneyDestockDataException {
         getStockTransactionCashMoneyDestockDao().saveCashMoneyDestockTransactionData(cashMoneyTransaction);
     }
